BRMS-76-better-logs: updated to fix review comments. moved common function in common util.
